refactor: evm reader organization
Also fix claim status checker
